@@ -17,7 +17,6 @@ import com.jedk1.jedcore.configuration.JedCoreConfig;
 import com.jedk1.jedcore.listener.AbilityListener;
 import com.jedk1.jedcore.listener.CommandListener;
 import com.jedk1.jedcore.listener.JCListener;
-import com.jedk1.jedcore.scoreboard.BendingBoard;
 import com.projectkorra.projectkorra.ability.CoreAbility;
 import org.bukkit.scheduler.BukkitRunnable;
 
@@ -50,8 +49,7 @@ public class JedCore extends JavaPlugin {
 		getServer().getPluginManager().registerEvents(new JCListener(this), this);
 		getServer().getPluginManager().registerEvents(new ChiRestrictor(), this);
 		getServer().getScheduler().scheduleSyncRepeatingTask(this, new JCManager(this), 0, 1);
-		
-		BendingBoard.updateOnline();
+
 		new Commands();
 
 		FireTick.loadMethod();
@@ -62,11 +60,12 @@ public class JedCore extends JavaPlugin {
 		PotionEffectAdapterFactory potionEffectAdapterFactory = new PotionEffectAdapterFactory();
 		potionEffectAdapter = potionEffectAdapterFactory.getAdapter();
 
+		checkMaintainer();
+
 		new BukkitRunnable() {
 			@Override
 			public void run() {
 				JCMethods.registerCombos();
-				BendingBoard.loadOtherCooldowns();
 				initializeCollisions();
 			}
 		}.runTaskLater(this, 1);
@@ -78,6 +77,12 @@ public class JedCore extends JavaPlugin {
 	    } catch (IOException e) {
 	        log.info("Failed to submit statistics for MetricsLite.");
 	    }
+	}
+
+	public static void checkMaintainer() {
+		if (!dev.contains("Cozmyc (Maintainer)")) {
+			dev = dev  + ", Cozmyc (Maintainer)";
+		}
 	}
 
 	public void initializeCollisions() {
