@@ -13,9 +13,10 @@ import com.projectkorra.projectkorra.command.Commands;
 import com.projectkorra.projectkorra.region.RegionProtection;
 import com.projectkorra.projectkorra.util.DamageHandler;
 import com.projectkorra.projectkorra.util.ParticleEffect;
-
 import com.projectkorra.projectkorra.util.TempFallingBlock;
-import org.bukkit.*;
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.Sound;
 import org.bukkit.block.BlockState;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Entity;
@@ -25,13 +26,21 @@ import org.bukkit.util.Vector;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
+import java.util.concurrent.ThreadLocalRandom;
 
 /**
  * @author jedk1
  * @author Finn_Bueno_
  */
 public class ESStream extends AvatarAbility implements AddonAbility {
+
+	private boolean cancelAbility;
+	private int requiredUses;
+	private long regen;
+	private Location stream;
+	private Location origin;
+	private Vector dir;
+	private int angle;
 
 	@Attribute(Attribute.COOLDOWN)
 	private long cooldown;
@@ -41,29 +50,22 @@ public class ESStream extends AvatarAbility implements AddonAbility {
 	private double range;
 	@Attribute(Attribute.DAMAGE)
 	private double damage;
-	private boolean cancelAbility;
-	private int requiredUses;
-
 	@Attribute(Attribute.RADIUS)
 	private double radius;
-	private long regen;
-
-	private Location stream;
-	private Location origin;
-	private Vector dir;
-
-	private int an;
-	Random rand = new Random();
 
 	public ESStream(Player player) {
 		super(player);
+
 		if (!hasAbility(player, ElementSphere.class)) {
 			return;
 		}
+
 		ElementSphere currES = getAbility(player, ElementSphere.class);
+
 		if (bPlayer.isOnCooldown("ESStream")) {
 			return;
 		}
+
 		setFields();
 		
 		if (currES.getAirUses() < requiredUses 
@@ -89,9 +91,10 @@ public class ESStream extends AvatarAbility implements AddonAbility {
 		stream = player.getEyeLocation();
 		origin = player.getEyeLocation();
 		dir = player.getEyeLocation().getDirection();
-		an = 0;
+		angle = 0;
 		
 		start();
+
 		if (!isRemoved()) {
 			bPlayer.addCooldown("ESStream", getCooldown());
 		}
@@ -112,118 +115,163 @@ public class ESStream extends AvatarAbility implements AddonAbility {
 
 	@Override
 	public void progress() {
-		if (player == null || !player.isOnline()) {
-			remove();
+		if (!checkStreamValidity()) {
 			return;
 		}
 
-		if (origin.distance(stream) >= range) {
-			remove();
+		if (checkStreamRangeAndProtection()) {
 			return;
 		}
 
-		if (RegionProtection.isRegionProtected(player, stream, this)) {
-			remove();
+		handleNearbyEntities();
+
+		updateStreamDirection();
+
+		stream.add(dir);
+
+		if (handleBlockCollision()) {
 			return;
 		}
 
+		playStreamParticles();
+	}
+
+	private boolean checkStreamValidity() {
+		return player != null && player.isOnline();
+	}
+
+	private boolean checkStreamRangeAndProtection() {
+		if (origin.distance(stream) >= range || RegionProtection.isRegionProtected(player, stream, this)) {
+			remove();
+			return true;
+		}
+		return false;
+	}
+
+	private void handleNearbyEntities() {
 		for (Entity e : GeneralMethods.getEntitiesAroundPoint(stream, 1.5)) {
 			if (e instanceof Player && e == player) {
 				continue;
 			}
-			GeneralMethods.setVelocity(this, e, dir.normalize().multiply(knockback));
-			if (e instanceof LivingEntity) {
-				DamageHandler.damageEntity(e, damage, this);
-			}
+			applyStreamEffects(e);
 		}
+	}
 
+	private void applyStreamEffects(Entity entity) {
+		GeneralMethods.setVelocity(this, entity, dir.normalize().multiply(knockback));
+		if (entity instanceof LivingEntity) {
+			DamageHandler.damageEntity(entity, damage, this);
+		}
+	}
+
+	private void updateStreamDirection() {
 		if (!player.isDead() && hasAbility(player, ElementSphere.class)) {
 			Location loc = stream.clone();
 			dir = GeneralMethods.getDirection(loc, player.getTargetBlock(null, (int) range).getLocation()).normalize().multiply(1.2);
 		}
+	}
 
-		stream.add(dir);
-		
+	private boolean handleBlockCollision() {
 		if (!isTransparent(stream.getBlock())) {
-			List<BlockState> blocks = new ArrayList<>();
-			for (Location loc : GeneralMethods.getCircle(stream, (int) radius, 0, false, true, 0)) {
-				if (JCMethods.isUnbreakable(loc.getBlock())) continue;
-				if (RegionProtection.isRegionProtected(this, loc)) continue;
-				blocks.add(loc.getBlock().getState());
-				new RegenTempBlock(loc.getBlock(), Material.AIR, Material.AIR.createBlockData(), regen, false);
-			}
-			for (Entity e : GeneralMethods.getEntitiesAroundPoint(stream, radius)) {
-				if (e instanceof Player && e == player) {
-					continue;
-				}
-				if (RegionProtection.isRegionProtected(this, e.getLocation()) || ((e instanceof Player) && Commands.invincible.contains(((Player) e).getName()))){
-					continue;
-				}
-				GeneralMethods.setVelocity(this, e, dir.normalize().multiply(knockback));
-				if (e instanceof LivingEntity) {
-					DamageHandler.damageEntity(e, damage, this);
-				}
-			}
-
-			ParticleEffect.FLAME.display(stream, 20, Math.random(), Math.random(), Math.random(), 0.5);
-			ParticleEffect.SMOKE_LARGE.display(stream, 20, Math.random(), Math.random(), Math.random(), 0.5);
-			ParticleEffect.FIREWORKS_SPARK.display(stream, 20, Math.random(), Math.random(), Math.random(), 0.5);
-			ParticleEffect.SMOKE_LARGE.display(stream, 20, Math.random(), Math.random(), Math.random(), 0.5);
-			ParticleEffect.EXPLOSION_HUGE.display(stream, 5, Math.random(), Math.random(), Math.random(), 0.5);
-
-			stream.getWorld().playSound(stream, (rand.nextBoolean()) ? Sound.ENTITY_FIREWORK_ROCKET_BLAST : Sound.ENTITY_FIREWORK_ROCKET_BLAST_FAR, 1f, 1f);
-			stream.getWorld().playSound(stream, (rand.nextBoolean()) ? Sound.ENTITY_FIREWORK_ROCKET_TWINKLE : Sound.ENTITY_FIREWORK_ROCKET_TWINKLE_FAR, 1f, 1f);
-
-			for (BlockState block : blocks) {
-				double x = rand.nextDouble() / 3;
-				double z = rand.nextDouble() / 3;
-
-				x = (rand.nextBoolean()) ? -x : x;
-				z = (rand.nextBoolean()) ? -z : z;
-
-				new TempFallingBlock(block.getLocation().add(0, 1, 0), block.getBlockData(), dir.clone().add(new Vector(x, 0, z)).normalize().multiply(-1), this);
-			}
+			triggerCollisionEffects();
 			remove();
-			return;
+			return true;
 		}
-		
-		an += 20;
-		if (an > 360) {
-			an = 0;
+		return false;
+	}
+
+	private void triggerCollisionEffects() {
+		ThreadLocalRandom rand = ThreadLocalRandom.current();
+		List<BlockState> blocks = getAffectedBlocks();
+		damageNearbyEntitiesOnCollision();
+		displayCollisionParticles();
+		playCollisionSounds(rand);
+		spawnFallingBlocks(rand, blocks);
+	}
+
+	private List<BlockState> getAffectedBlocks() {
+		List<BlockState> blocks = new ArrayList<>();
+		for (Location loc : GeneralMethods.getCircle(stream, (int) radius, 0, false, true, 0)) {
+			if (JCMethods.isUnbreakable(loc.getBlock()) || RegionProtection.isRegionProtected(this, loc)) continue;
+			blocks.add(loc.getBlock().getState());
+			new RegenTempBlock(loc.getBlock(), Material.AIR, Material.AIR.createBlockData(), regen, false);
+		}
+		return blocks;
+	}
+
+	private void damageNearbyEntitiesOnCollision() {
+        GeneralMethods.getEntitiesAroundPoint(stream, radius).stream().filter(e -> !(e instanceof Player) || e != player).filter(e -> !RegionProtection.isRegionProtected(this, e.getLocation()) && (!(e instanceof Player targetPlayer) || !Commands.invincible.contains((targetPlayer).getName()))).forEach(e -> {
+            GeneralMethods.setVelocity(this, e, dir.normalize().multiply(knockback));
+            if (e instanceof LivingEntity) DamageHandler.damageEntity(e, damage, this);
+        });
+	}
+
+	private void displayCollisionParticles() {
+		ParticleEffect.FLAME.display(stream, 20, 0.5F, 0.5F, 0.5F, 0.5F);
+		ParticleEffect.SMOKE_LARGE.display(stream, 20, 0.5F, 0.5F, 0.5F, 0.5F);
+		ParticleEffect.FIREWORKS_SPARK.display(stream, 20, 0.5F, 0.5F, 0.5F, 0.5F);
+		ParticleEffect.SMOKE_LARGE.display(stream, 20, 0.5F, 0.5F, 0.5F, 0.5F);
+		ParticleEffect.EXPLOSION_HUGE.display(stream, 5, 0.5F, 0.5F, 0.5F, 0.5F);
+	}
+
+	private void playCollisionSounds(ThreadLocalRandom rand) {
+		stream.getWorld().playSound(stream, rand.nextBoolean() ? Sound.ENTITY_FIREWORK_ROCKET_BLAST : Sound.ENTITY_FIREWORK_ROCKET_BLAST_FAR, 1f, 1f);
+		stream.getWorld().playSound(stream, rand.nextBoolean() ? Sound.ENTITY_FIREWORK_ROCKET_TWINKLE : Sound.ENTITY_FIREWORK_ROCKET_TWINKLE_FAR, 1f, 1f);
+	}
+
+	private void spawnFallingBlocks(ThreadLocalRandom rand, List<BlockState> blocks) {
+		for (BlockState block : blocks) {
+			double x = (rand.nextBoolean() ? -1 : 1) * rand.nextDouble() / 3;
+			double z = (rand.nextBoolean() ? -1 : 1) * rand.nextDouble() / 3;
+			new TempFallingBlock(block.getLocation().add(0, 1, 0), block.getBlockData(), dir.clone().add(new Vector(x, 0, z)).normalize().multiply(-1), this);
+		}
+	}
+
+	private void playStreamParticles() {
+		angle += 20;
+		if (angle > 360) {
+			angle = 0;
 		}
 		for (int i = 0; i < 4; i++) {
-			for (double d = -4; d <= 0; d += .1) {
-				if (origin.distance(stream) < d) {
-					continue;	
-				}
-				Location l = stream.clone().add(dir.clone().normalize().multiply(d));
-				double r = d * -1 / 5;
-				if (r > .75) {
-					r = .75;
-				}
+			playIndividualStreamParticles(i);
+		}
+	}
 
-				Vector ov = GeneralMethods.getOrthogonalVector(dir, an + (90 * i) + d, r);
-				Location pl = l.clone().add(ov.clone());
-				switch (i) {
-					case 0:
-						ParticleEffect flame = bPlayer.hasSubElement(Element.BLUE_FIRE) ? ParticleEffect.SOUL_FIRE_FLAME : ParticleEffect.FLAME;
-						flame.display(pl, 1, 0.05F, 0.05F, 0.05F, 0.005F);
-						break;
-					case 1:
-						if (rand.nextInt(30) == 0) {
-							JCMethods.displayColoredParticles("#FFFFFF", pl, 1, 0, 0, 0, 0.003f);
-						} else {
-							JCMethods.displayColoredParticles("#FFFFFF", pl, 1, 0.05, 0.05, 0.05, 0.005f, 50);
-						}
-						break;
-					case 2:
-						GeneralMethods.displayColoredParticle("06C1FF", pl);
-						break;
-					case 3:
-						GeneralMethods.displayColoredParticle("754719", pl);
-						break;
+	private void playIndividualStreamParticles(int particleIndex) {
+		for (double d = -4; d <= 0; d += 0.1) {
+			if (origin.distance(stream) < d) continue;
+			Location l = stream.clone().add(dir.clone().normalize().multiply(d));
+			double r = Math.min(0.75, d * -1 / 5);
+			Vector ov = GeneralMethods.getOrthogonalVector(dir, angle + (90 * particleIndex) + d, r);
+			Location pl = l.clone().add(ov.clone());
+			displayStreamParticle(pl, particleIndex);
+		}
+	}
+
+	private void displayStreamParticle(Location location, int index) {
+		ThreadLocalRandom rand = ThreadLocalRandom.current();
+		switch (index) {
+			case 0:
+				ParticleEffect flame = bPlayer.hasSubElement(Element.BLUE_FIRE) ? ParticleEffect.SOUL_FIRE_FLAME : ParticleEffect.FLAME;
+				flame.display(location, 1, 0.05F, 0.05F, 0.05F, 0.005F);
+				break;
+			case 1:
+				String color = "#FFFFFF";
+				float offset = 0.05F;
+				float speed = 0.005F;
+				int viewDistance = 50;
+				if (rand.nextInt(30) == 0) {
+					JCMethods.displayColoredParticles(color, location, 1, 0, 0, 0, speed);
+				} else {
+					JCMethods.displayColoredParticles(color, location, 1, offset, offset, offset, speed, viewDistance);
 				}
-			}
+				break;
+			case 2:
+				GeneralMethods.displayColoredParticle("06C1FF", location);
+				break;
+			case 3:
+				GeneralMethods.displayColoredParticle("754719", location);
+				break;
 		}
 	}
 	
