@@ -5,13 +5,16 @@ import com.jedk1.jedcore.JedCore;
 import com.jedk1.jedcore.collision.CollisionDetector;
 import com.jedk1.jedcore.collision.Sphere;
 import com.jedk1.jedcore.configuration.JedCoreConfig;
-import com.jedk1.jedcore.policies.removal.*;
+import com.jedk1.jedcore.policies.removal.CannotBendRemovalPolicy;
+import com.jedk1.jedcore.policies.removal.CompositeRemovalPolicy;
+import com.jedk1.jedcore.policies.removal.IsDeadRemovalPolicy;
+import com.jedk1.jedcore.policies.removal.IsOfflineRemovalPolicy;
+import com.jedk1.jedcore.policies.removal.SwappedSlotsRemovalPolicy;
 import com.jedk1.jedcore.util.FireTick;
-import com.jedk1.jedcore.util.LightManager;
 import com.jedk1.jedcore.util.MaterialUtil;
 import com.jedk1.jedcore.util.RegenTempBlock;
-import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.Element.SubElement;
+import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ability.AddonAbility;
 import com.projectkorra.projectkorra.ability.AirAbility;
 import com.projectkorra.projectkorra.ability.CombustionAbility;
@@ -22,10 +25,12 @@ import com.projectkorra.projectkorra.region.RegionProtection;
 import com.projectkorra.projectkorra.util.DamageHandler;
 import com.projectkorra.projectkorra.util.ParticleEffect;
 import com.projectkorra.projectkorra.util.TempBlock;
-
+import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.NamespacedKey;
 import org.bukkit.Sound;
+import org.bukkit.Tag;
 import org.bukkit.block.Block;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Entity;
@@ -33,9 +38,11 @@ import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
+import java.util.stream.Collectors;
 
 public class Combustion extends CombustionAbility implements AddonAbility {
 
@@ -44,6 +51,8 @@ public class Combustion extends CombustionAbility implements AddonAbility {
 	@Attribute(Attribute.COOLDOWN)
 	private long cooldown;
 	private CompositeRemovalPolicy removalPolicy;
+
+	private ArrayList<String> skipMaterials; // use a configured list of blocks to skip through
 
 	public Combustion(Player player) {
 		super(player);
@@ -81,6 +90,8 @@ public class Combustion extends CombustionAbility implements AddonAbility {
 		);
 
 		this.removalPolicy.load(config, "Abilities.Fire.Combustion");
+
+		this.skipMaterials = loadSkipMaterials();
 	}
 
 	@Override
@@ -169,6 +180,35 @@ public class Combustion extends CombustionAbility implements AddonAbility {
 	public boolean isEnabled() {
 		ConfigurationSection config = JedCoreConfig.getConfig(this.player);
 		return config.getBoolean("Abilities.Fire.Combustion.Enabled");
+	}
+
+	private ArrayList<String> loadSkipMaterials() {
+		ConfigurationSection config = JedCoreConfig.getConfig(this.player);
+
+		ArrayList<String> skipList = new ArrayList<>();
+
+		if (config.contains("Abilities.Fire.Combustion.SkipMaterials")) {
+			List<String> configuredSkipList = config.getStringList("Abilities.Fire.Combustion.SkipMaterials");
+
+			for (String entry : configuredSkipList) {
+				if (entry.startsWith("#")) {
+					String tagName = entry.substring(1).toLowerCase();
+
+					NamespacedKey tagKey = NamespacedKey.minecraft(tagName);
+					Tag<Material> materialTag = Bukkit.getTag(Tag.REGISTRY_BLOCKS, tagKey, Material.class);
+
+					if (materialTag != null) {
+						skipList.addAll(materialTag.getValues().stream()
+								.map(material -> material.name().toLowerCase())
+								.collect(Collectors.toList()));
+					}
+				} else {
+					skipList.add(entry.toLowerCase());
+				}
+			}
+		}
+
+		return skipList;
 	}
 
 	private interface State {
@@ -328,8 +368,15 @@ public class Combustion extends CombustionAbility implements AddonAbility {
 				}
 
 				if (!MaterialUtil.isTransparent(location.getBlock()) || isWater(location.getBlock())) {
-					state = new CombustState(location);
-					return;
+					Material blockMaterial = location.getBlock().getType();
+					String blockMaterialName = blockMaterial.name().toLowerCase();
+
+					boolean shouldSkip = skipMaterials.contains(blockMaterialName);
+
+					if (!shouldSkip) {
+						state = new CombustState(location);
+						return;
+					}
 				}
 
 				direction = player.getEyeLocation().getDirection().normalize();
