@@ -1,15 +1,14 @@
 package com.jedk1.jedcore.command;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.UUID;
-
+import com.jedk1.jedcore.JedCore;
+import com.projectkorra.projectkorra.command.PKCommand;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import com.jedk1.jedcore.JedCore;
-import com.projectkorra.projectkorra.command.PKCommand;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
 
 public class JedCoreCommand extends PKCommand {
 	private static final String DOWNLOAD_URL = "https://github.com/CozmycDev/JedCore";
