@@ -63,17 +63,20 @@ import com.projectkorra.projectkorra.earthbending.EarthArmor;
 import com.projectkorra.projectkorra.earthbending.lava.LavaFlow;
 import com.projectkorra.projectkorra.firebending.FireJet;
 import com.projectkorra.projectkorra.util.MovementHandler;
+import com.projectkorra.projectkorra.util.ParticleEffect;
 import com.projectkorra.projectkorra.waterbending.blood.Bloodbending;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.entity.Arrow;
 import org.bukkit.entity.EntityType;
+import org.bukkit.entity.FallingBlock;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
+import org.bukkit.event.entity.EntityChangeBlockEvent;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.inventory.InventoryAction;
 import org.bukkit.event.inventory.InventoryClickEvent;
@@ -154,6 +157,21 @@ public class AbilityListener implements Listener {
 	@EventHandler(priority = EventPriority.LOWEST)
 	public void onPlayerDropItemEvent(PlayerDropItemEvent event) {
 		recentlyDropped.add(event.getPlayer().getUniqueId());
+
+		// FirePunch item burn effect
+		Player player = event.getPlayer();
+		FirePunch fp = FirePunch.getAbility(player, FirePunch.class);
+		boolean burnEnabled = com.jedk1.jedcore.configuration.JedCoreConfig.getConfig((Player)null).getBoolean("Abilities.Fire.FirePunch.BurnsDroppedItems", true);
+		if (fp != null && burnEnabled) {
+			var item = event.getItemDrop();
+			var loc = fp.getRightHandPos();
+			ParticleEffect.LAVA.display(loc, 5, 0.1, 0.1, 0.1, 0.05);
+			ParticleEffect.FLAME.display(loc, 3, 0.1, 0.1, 0.1, 0.01);
+			ParticleEffect.SMOKE_NORMAL.display(loc, 2, 0.05, 0.05, 0.05, 0.01);
+			loc.getWorld().playSound(loc, Sound.BLOCK_FIRE_EXTINGUISH, 0.3f, 1.4f);
+			item.remove();
+		}
+
 		Bukkit.getScheduler().runTaskLater(ProjectKorra.plugin, () -> recentlyDropped.remove(event.getPlayer().getUniqueId()), 2L);
 	}
 
@@ -262,7 +280,10 @@ public class AbilityListener implements Listener {
 					new LavaFlux(player);
 				}
 				if (abilClass.equals(LavaThrow.class)) {
-					new LavaThrow(player);
+					LavaThrow lt = CoreAbility.getAbility(player, LavaThrow.class);
+					if (lt != null) {
+						lt.createBlast();
+					}
 				}
 				if (abilClass.equals(MetalFragments.class)) {
 					MetalFragments.shootFragment(player);
@@ -393,6 +414,10 @@ public class AbilityListener implements Listener {
 				if (abilClass.equals(FireShots.class)) {
 					FireShots.swapHands(player);
 				}
+
+				if (abilClass.equals(FirePunch.class)) {
+					FirePunch.swapHands(player);
+				}
 			}
 		}
 
@@ -455,6 +480,12 @@ public class AbilityListener implements Listener {
 				if (abilClass.equals(Crevice.class)) {
 					Crevice.closeCrevice(player);
 				}
+				if (abilClass.equals(LavaThrow.class)) {
+					LavaThrow lt = CoreAbility.getAbility(player, LavaThrow.class);
+					if (lt == null) {
+						new LavaThrow(player);
+					}
+				}
 			}
 
 			if (coreAbil instanceof FireAbility && bPlayer.isElementToggled(Element.FIRE)) {
@@ -504,6 +535,7 @@ public class AbilityListener implements Listener {
 				}
 				if (abilClass.equals(IceClaws.class)) {
 					new IceClaws(player);
+					IceClaws.swapHands(player);
 				}
 				if (abilClass.equals(IceWall.class)) {
 					new IceWall(player);
@@ -528,9 +560,10 @@ public class AbilityListener implements Listener {
 	public void onArrowHit(EntityDamageByEntityEvent event) {
 		if (event.getDamager().getType() == EntityType.ARROW) {
 			Arrow arrow = (Arrow) event.getDamager();
-			if (arrow.getShooter() instanceof Player && arrow.hasMetadata("daggerthrow")){
+			if (arrow.getShooter() instanceof Player && arrow.hasMetadata("daggerthrow")) {
 				Player player = (Player) arrow.getShooter();
-				player.playSound(player.getLocation(), Sound.ENTITY_ARROW_HIT_PLAYER, 1f, 1);
+				event.setDamage(0.0D);
+				player.playSound(player.getLocation(), Sound.ENTITY_ARROW_HIT_PLAYER, 1.0f, 1.0f);
 			}
 		}
 	}
@@ -539,6 +572,13 @@ public class AbilityListener implements Listener {
 	public void onPlayerInteraction(PlayerInteractEvent event) {
 		if (event.getAction() == Action.RIGHT_CLICK_BLOCK) {
 			MetalFragments.shootFragment(event.getPlayer());
+		}
+	}
+
+	@EventHandler
+	public void onChange(EntityChangeBlockEvent event) {
+		if (event.getEntity() instanceof FallingBlock) {
+			if (event.getEntity().hasMetadata("magmablast")) event.setCancelled(true);
 		}
 	}
 }
